apply rules
