+++ tests/cloudsqlmssql/cloud_sql_mssql_integration_test.go
@@ -17,94 +17,17 @@ package cloudsqlmssql
 import (
 	"context"
 	"database/sql"
-	"fmt"
-	"net/url"
-	"os"
 	"regexp"
-	"slices"
 	"strings"
 	"testing"
 	"time"
 
-	"cloud.google.com/go/cloudsqlconn"
-	"cloud.google.com/go/cloudsqlconn/sqlserver/mssql"
 	"github.com/google/uuid"
 	"github.com/googleapis/mcp-toolbox/internal/testutils"
 	"github.com/googleapis/mcp-toolbox/tests"
 )
 
-var (
-	CloudSQLMSSQLSourceType = "cloud-sql-mssql"
-	CloudSQLMSSQLToolType   = "mssql-sql"
-	CloudSQLMSSQLProject    = os.Getenv("CLOUD_SQL_MSSQL_PROJECT")
-	CloudSQLMSSQLRegion     = os.Getenv("CLOUD_SQL_MSSQL_REGION")
-	CloudSQLMSSQLInstance   = os.Getenv("CLOUD_SQL_MSSQL_INSTANCE")
-	CloudSQLMSSQLDatabase   = os.Getenv("CLOUD_SQL_MSSQL_DATABASE")
-	CloudSQLMSSQLUser       = os.Getenv("CLOUD_SQL_MSSQL_USER")
-	CloudSQLMSSQLPass       = os.Getenv("CLOUD_SQL_MSSQL_PASS")
-)
 
-func getCloudSQLMSSQLVars(t *testing.T) map[string]any {
-	switch "" {
-	case CloudSQLMSSQLProject:
-		t.Fatal("'CLOUD_SQL_MSSQL_PROJECT' not set")
-	case CloudSQLMSSQLRegion:
-		t.Fatal("'CLOUD_SQL_MSSQL_REGION' not set")
-	case CloudSQLMSSQLInstance:
-		t.Fatal("'CLOUD_SQL_MSSQL_INSTANCE' not set")
-	case CloudSQLMSSQLDatabase:
-		t.Fatal("'CLOUD_SQL_MSSQL_DATABASE' not set")
-	case CloudSQLMSSQLUser:
-		t.Fatal("'CLOUD_SQL_MSSQL_USER' not set")
-	case CloudSQLMSSQLPass:
-		t.Fatal("'CLOUD_SQL_MSSQL_PASS' not set")
-	}
-
-	return map[string]any{
-		"type":     CloudSQLMSSQLSourceType,
-		"project":  CloudSQLMSSQLProject,
-		"instance": CloudSQLMSSQLInstance,
-		"region":   CloudSQLMSSQLRegion,
-		"database": CloudSQLMSSQLDatabase,
-		"user":     CloudSQLMSSQLUser,
-		"password": CloudSQLMSSQLPass,
-	}
-}
-
-// Copied over from cloud_sql_mssql.go
-func initCloudSQLMSSQLConnection(project, region, instance, ipType, user, pass, dbname string) (*sql.DB, error) {
-	// Create dsn
-	query := fmt.Sprintf("database=%s&cloudsql=%s:%s:%s", dbname, project, region, instance)
-	url := &url.URL{
-		Scheme:   "sqlserver",
-		User:     url.UserPassword(user, pass),
-		RawQuery: query,
-	}
-
-	// Get dial options
-	dialOpts, err := tests.GetCloudSQLDialOpts(ipType)
-	if err != nil {
-		return nil, err
-	}
-
-	// Register sql server driver
-	if !slices.Contains(sql.Drivers(), "cloudsql-sqlserver-driver") {
-		_, err := mssql.RegisterDriver("cloudsql-sqlserver-driver", cloudsqlconn.WithDefaultDialOptions(dialOpts...))
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	// Open database connection
-	db, err := sql.Open(
-		"cloudsql-sqlserver-driver",
-		url.String(),
-	)
-	if err != nil {
-		return nil, err
-	}
-	return db, nil
-}
 
 func TestCloudSQLMSSQLToolEndpoints(t *testing.T) {
 	sourceConfig := getCloudSQLMSSQLVars(t)
@@ -169,32 +92,4 @@ func TestCloudSQLMSSQLToolEndpoints(t *testing.T) {
 
 	// Run specific MSSQL tool tests
 	tests.RunMSSQLListTablesTest(t, tableNameParam, tableNameAuth)
-}
-
-// Test connection with different IP type
-func TestCloudSQLMSSQLIpConnection(t *testing.T) {
-	sourceConfig := getCloudSQLMSSQLVars(t)
-
-	tcs := []struct {
-		name   string
-		ipType string
-	}{
-		{
-			name:   "public ip",
-			ipType: "public",
-		},
-		{
-			name:   "private ip",
-			ipType: "private",
-		},
-	}
-	for _, tc := range tcs {
-		t.Run(tc.name, func(t *testing.T) {
-			sourceConfig["ipType"] = tc.ipType
-			err := tests.RunSourceConnectionTest(t, sourceConfig, CloudSQLMSSQLToolType)
-			if err != nil {
-				t.Fatalf("Connection test failure: %s", err)
-			}
-		})
-	}
 }
